requirements.txt
apify-client
python-dotenv
crewai-tools
crewai
crewai[tools]
langchain_openai
